Revert "Add Coin game"
