+++ pax/conf/config.yaml
@@ -14,13 +14,13 @@ save_dir: "./exp/${wandb.group}/${wandb.name}"
 debug: False
 
 # Agents  
-agent1: 'PPO'
-agent2: 'PPO'
+agent1: 'PPO_memory'
+agent2: 'Naive'
 
 # Environment
 env_id: ipd
 game: ipd 
-env_type: coin_game
+env_type: meta
 env_discount: 0.96
 payoff: [[-1, -1], [-3, 0], [0, -3], [-2, -2]]
 
@@ -29,12 +29,13 @@ evo: False
 
 # Training hyperparameters
 # env_batch_size = num_envs * num_opponents
-num_envs: 100
-num_opps: 1
-num_steps: 16 # total number of steps per episode (this is num_inner_steps * num_outer_steps)
-num_inner_steps: 16 # number of inner steps
-total_timesteps: 1.0e8
-eval_every: 1.0e8 # timesteps 
+popsize: 20
+num_envs: 2
+num_opps: 60
+num_steps: 100 # total number of steps per episode (this is num_inner_steps * num_outer_steps)
+num_inner_steps: 25 # number of inner steps (only for MetaFinite Env)
+total_timesteps: 1.0e9
+eval_every: 1.0e9 # timesteps 
 
 # Useful information
 # num_episodes = total_timesteps / (num_steps * num_envs) 
@@ -43,7 +44,7 @@ eval_every: 1.0e8 # timesteps
 
 # PPO 
 ppo:
-  num_minibatches: 8
+  num_minibatches: 10
   num_epochs: 4 
   gamma: 0.96
   gae_lambda: 0.95
@@ -59,7 +60,6 @@ ppo:
   learning_rate: 3e-4
   adam_epsilon: 1e-5
   with_memory: True
-  hidden: 25
 
 # Naive Learner parameters 
 naive:
@@ -93,6 +93,17 @@ es:
 wandb:
   entity: "ucl-dark"
   project: ipd
-  group: ??
+  group: 'MARL^2-${agent1}-vs-${agent2}'
   name: run-seed-${seed}
-  log: True
+  log: False
+
+# DQN agent parameters
+dqn: 
+  batch_size: 256
+  discount: 0.99
+  learning_rate: 1e-2
+  epsilon: 0.5
+  replay_capacity: 100000
+  min_replay_size: 1000
+  sgd_period: 1 
+  target_update_period: 4
