--- pax/conf/experiment/cg/sanity.yaml
@@ -1,59 +1,0 @@
-# @package _global_
-
-# Agents  
-agent1: 'PPO_memory'
-agent2: 'Random'
-
-# Environment
-env_id: coin_game
-game: ipd 
-env_type: coin_game
-env_discount: 0.96
-payoff: [[-1, -1], [-3, 0], [0, -3], [-2, -2]]
-
-# Training hyperparameters
-
-# env_batch_size = num_envs * num_opponents
-num_envs: 1000
-num_opps: 1
-num_steps: 16 # total number of steps per episode (this is num_inner_steps * num_outer_steps)
-num_inner_steps: 16 # number of inner steps
-total_timesteps: 4.0e8
-eval_every: 4.0e8 # timesteps 
-
-# Useful information
-# num_episodes = total_timesteps / (num_steps * num_envs) 
-# num_updates = num_episodes / eval_every
-# train_batch_size = num_envs * num_opponents * num_steps
-
-# PPO agent parameters
-ppo:
-  num_minibatches: 8
-  num_epochs: 2 
-  gamma: 0.96
-  gae_lambda: 0.95
-  ppo_clipping_epsilon: 0.2
-  value_coeff: 0.5
-  clip_value: True
-  max_gradient_norm: 0.5
-  anneal_entropy: True
-  entropy_coeff_start: 0.01
-  entropy_coeff_horizon: 2.0e8
-  entropy_coeff_end: 0.005
-  lr_scheduling: True
-  learning_rate: 3e-4
-  adam_epsilon: 1e-5
-  with_memory: True
-  output_channels: 16
-  kernel_shape: [3, 3]
-  with_cnn: True
-  separate: True # only works with CNN
-  hidden: 25
-
-# Logging setup
-wandb:
-  entity: "ucl-dark"
-  project: cg
-  group: 'sanity-${agent1}-vs-${agent2}-parity'
-  name: run-seed-${seed}
-  log: True
