+++ pax/conf/experiment/ipd/ppo.yaml
@@ -7,7 +7,7 @@ agent2: 'TitForTat'
 # Environment
 env_id: ipd
 game: ipd 
-env_type: sequential
+env_type: finite
 env_discount: 0.99
 payoff: [[-1,-1], [-3,0], [0,-3], [-2,-2]]
 
@@ -16,7 +16,7 @@ num_envs: 100
 num_opps: 1
 num_steps: 150 # number of steps per episode
 total_timesteps: 4_000_000
-eval_every: 4_000_000 # timesteps
+eval_every: 2_000_000 # timesteps
 
 # Useful information
 # num_episodes = total_timesteps / num_steps
@@ -42,7 +42,6 @@ ppo:
   learning_rate: 2.5e-2
   adam_epsilon: 1e-5
   with_memory: False
-  with_cnn: False
 
 # Logging setup
 wandb:
