+++ pax/env_inner.py
@@ -230,159 +230,3 @@ if __name__ == "__main__":
     env = SequentialMatrixGame(
         5, 1, jnp.array([[2, 2], [0, 3], [3, 0], [1, 1]])
     )
-
-
-class InfiniteMatrixGame(Environment):
-    def __init__(
-        self,
-        num_envs: int,
-        payoff: list,
-        episode_length: int,
-        gamma: float,
-        seed: int,
-    ) -> None:
-        self.payoff = jnp.array(payoff)
-        payout_mat_1 = jnp.array([[r[0] for r in self.payoff]])
-        payout_mat_2 = jnp.array([[r[1] for r in self.payoff]])
-
-        def _step(actions, state):
-            theta1, theta2 = actions
-            (t, rng) = state
-            rng, _ = jax.random.split(rng, 2)
-            theta1, theta2 = jax.nn.sigmoid(theta1), jax.nn.sigmoid(theta2)
-            obs1 = jnp.concatenate([theta1, theta2])
-            obs2 = jnp.concatenate([theta2, theta1])
-
-            _th2 = jnp.array(
-                [theta2[0], theta2[2], theta2[1], theta2[3], theta2[4]]
-            )
-
-            p_1_0 = theta1[4:5]
-            p_2_0 = _th2[4:5]
-            p = jnp.concatenate(
-                [
-                    p_1_0 * p_2_0,
-                    p_1_0 * (1 - p_2_0),
-                    (1 - p_1_0) * p_2_0,
-                    (1 - p_1_0) * (1 - p_2_0),
-                ]
-            )
-            p_1 = jnp.reshape(theta1[0:4], (4, 1))
-            p_2 = jnp.reshape(_th2[0:4], (4, 1))
-            P = jnp.concatenate(
-                [
-                    p_1 * p_2,
-                    p_1 * (1 - p_2),
-                    (1 - p_1) * p_2,
-                    (1 - p_1) * (1 - p_2),
-                ],
-                axis=1,
-            )
-            M = jnp.matmul(p, jnp.linalg.inv(jnp.eye(4) - gamma * P))
-            L_1 = jnp.matmul(M, jnp.reshape(payout_mat_1, (4, 1)))
-            L_2 = jnp.matmul(M, jnp.reshape(payout_mat_2, (4, 1)))
-            return (L_1.sum(), L_2.sum(), obs1, obs2, M), (t + 1, rng)
-
-        self._jit_step = jax.jit(jax.vmap(_step, (0, None), (0, None)))
-        self.gamma = gamma
-
-        self.num_envs = num_envs
-        self.episode_length = episode_length
-        self._num_steps = 0
-        self._reset_next_step = True
-
-        self.state = (0.0, jax.random.PRNGKey(seed=seed))
-        self.num_trials = 1
-        self.inner_episode_length = episode_length
-
-        # for runner
-        self.num_trials = 1
-        self.inner_episode_length = episode_length
-        self.batch_step = jax.jit(
-            jax.vmap(self.runner_step, (0, None), (0, None))
-        )
-
-    def step(
-        self, actions: Tuple[jnp.ndarray, jnp.ndarray]
-    ) -> Tuple[TimeStep, TimeStep]:
-        """
-        takes a tuple of batched policies and produce value functions from infinite game
-        policy of form [B, 5]
-        """
-        if self._reset_next_step:
-            return self.reset()
-
-        action_1, action_2 = actions
-        self._num_steps += 1
-        assert action_1.shape == action_2.shape
-        assert action_1.shape == (self.num_envs, 5)
-
-        outputs, self.state = self._jit_step(actions, self.state)
-        r1, r2, obs1, obs2, _ = outputs
-        r1, r2 = (1 - self.gamma) * r1, (1 - self.gamma) * r2
-
-        if self._num_steps == self.episode_length:
-            self._reset_next_step = True
-            return termination(reward=r1, observation=obs1), termination(
-                reward=r2, observation=obs2
-            )
-        return transition(reward=r1, observation=obs1), transition(
-            reward=r2, observation=obs2
-        )
-
-    def runner_step(
-        self,
-        actions: Tuple[jnp.ndarray, jnp.ndarray],
-        env_state: Tuple[float, float],
-    ) -> Tuple[Tuple[TimeStep, TimeStep], Tuple[float, float]]:
-
-        (r1, r2, obs1, obs2, _), env_state = self._jit_step(actions, env_state)
-        r1, r2 = (1 - self.gamma) * r1, (1 - self.gamma) * r2
-        return (
-            TimeStep(
-                jnp.ones_like(r1, dtype=int),
-                reward=r1,
-                discount=jnp.zeros_like(r1, dtype=int),
-                observation=obs1,
-            ),
-            TimeStep(
-                jnp.ones_like(r1, dtype=int),
-                reward=r2,
-                discount=jnp.zeros_like(r1, dtype=int),
-                observation=obs2,
-            ),
-        ), env_state
-
-    def observation_spec(self) -> specs.DiscreteArray:
-        """Returns the observation spec."""
-        return specs.DiscreteArray(num_values=10, name="previous policy")
-
-    def action_spec(self) -> specs.DiscreteArray:
-        """Returns the action spec."""
-        return specs.BoundedArray(
-            shape=(self.num_envs, 5),
-            minimum=0,
-            maximum=1,
-            name="action",
-            dtype=float,
-        )
-
-    def runner_reset(self, ndims, rng):
-        """Pure version of reset"""
-        key, _ = jax.random.split(rng)
-        new_state = (0, key)
-        discount = jnp.zeros(ndims, dtype=int)
-        step_type = jnp.zeros(ndims, dtype=int)
-        obs = jax.nn.sigmoid(jax.random.uniform(key, ndims + (10,)))
-        return (
-            TimeStep(step_type, jnp.zeros(ndims), discount, obs),
-            TimeStep(step_type, jnp.zeros(ndims), discount, obs),
-        ), (new_state)
-
-    def reset(self) -> Tuple[TimeStep, TimeStep]:
-        """Returns the first `TimeStep` of a new episode."""
-        self._reset_next_step = False
-        self._num_steps = 0
-        t_init, self.state = self.runner_reset((self.num_envs,), self.state[1])
-        self.key = self.state[1]
-        return t_init
