+++ pax/naive/network.py
@@ -4,6 +4,8 @@ import distrax
 import haiku as hk
 import jax.numpy as jnp
 
+from pax import utils
+
 
 class CategoricalValueHead(hk.Module):
     """Network head that produces a categorical distribution and value."""
