+++ pax/naive_exact.py
@@ -4,7 +4,7 @@ import jax
 import jax.numpy as jnp
 from dm_env import TimeStep
 
-from pax.env_inner import InfiniteMatrixGame
+from pax.env_meta import InfiniteMatrixGame
 
 
 class TrainingState(NamedTuple):
