+++ pax/runner_evo.py
@@ -11,7 +11,7 @@ import wandb
 
 # TODO: import when evosax library is updated
 # from evosax.utils import ESLog
-from pax.watchers import ESLog, ipd_visitation
+from pax.watchers import ESLog
 
 MAX_WANDB_CALLS = 1000
 
@@ -49,7 +49,30 @@ class EvoRunner:
         self.top_k = args.top_k
         self.train_steps = 0
         self.train_episodes = 0
-        self.ipd_stats = jax.jit(ipd_visitation)
+
+        def _state_visitation(traj: Sample, final_t: TimeStep) -> List:
+            # obs [num_outer_steps, num_inner_steps, num_opps, num_envs, ...]
+            # final_t [num_opps, num_envs, ...]
+            num_timesteps = (
+                traj.observations.shape[0] * traj.observations.shape[1]
+            )
+            # obs = [0, 1, 2, 3, 4], a = [0, 1]
+            # combine = [0, .... 9]
+            state_actions = (
+                2 * jnp.argmax(traj.observations, axis=-1) + traj.actions
+            )
+            state_actions = jnp.reshape(
+                state_actions,
+                (num_timesteps,) + state_actions.shape[2:],
+            )
+            # assume final step taken is cooperate
+            final_obs = jax.lax.expand_dims(
+                2 * jnp.argmax(final_t.observation, axis=-1), [0]
+            )
+            state_actions = jnp.append(state_actions, final_obs, axis=0)
+            return jnp.bincount(state_actions.flatten(), length=10)
+
+        self.state_visitation = jax.jit(_state_visitation)
 
     def train_loop(self, env, agents, num_episodes, watchers):
         """Run training of agents in environment"""
@@ -256,13 +279,12 @@ class EvoRunner:
             final_t1 = vals[0]._replace(
                 step_type=2 * jnp.ones_like(vals[0].step_type)
             )
+            visits = self.state_visitation(traj_1, final_t1)
+            states = visits.reshape((int(visits.shape[0] / 2), 2)).sum(axis=1)
+            state_freq = states / states.sum()
+            action_probs = visits[::2] / states
 
             if gen % log_interval == 0:
-                env_stats = jax.tree_util.tree_map(
-                    lambda x: x.item(), self.ipd_stats(traj_1, final_t1)
-                )
-
-                print(env_stats)
                 print(f"Generation: {gen}")
                 print(
                     "--------------------------------------------------------------------------"
@@ -270,8 +292,8 @@ class EvoRunner:
                 print(
                     f"Fitness: {fitness.mean()} | Other Fitness: {other_fitness.mean()}"
                 )
-                print(f"Env Visitation: {list(env_stats.values())[5:]}")
-                print(f"Cooperation Frequency: {list(env_stats.values())[:5]}")
+                print(f"State Visitation: {states}")
+                print(f"Cooperation Frequency: {action_probs}")
                 print(
                     "--------------------------------------------------------------------------"
                 )
@@ -292,6 +314,7 @@ class EvoRunner:
             self.generations += 1
 
             if watchers:
+
                 wandb_log = {
                     "generations": self.generations,
                     "train/fitness/player_1": float(fitness.mean()),
@@ -301,6 +324,16 @@ class EvoRunner:
                     "train/fitness/top_gen_mean": log["log_top_gen_mean"][gen],
                     "train/fitness/top_gen_std": log["log_top_gen_std"][gen],
                     "train/fitness/gen_std": log["log_gen_std"][gen],
+                    "train/state_visitation/CC": state_freq[0],
+                    "train/state_visitation/CD": state_freq[1],
+                    "train/state_visitation/DC": state_freq[2],
+                    "train/state_visitation/DD": state_freq[3],
+                    "train/state_visitation/START": state_freq[4],
+                    "train/cooperation_probability/CC": action_probs[0],
+                    "train/cooperation_probability/CD": action_probs[1],
+                    "train/cooperation_probability/DC": action_probs[2],
+                    "train/cooperation_probability/DD": action_probs[3],
+                    "train/cooperation_probability/START": action_probs[4],
                     "train/time/minutes": float(
                         (time.time() - self.start_time) / 60
                     ),
@@ -308,7 +341,7 @@ class EvoRunner:
                         (time.time() - self.start_time)
                     ),
                 }
-                wandb_log = wandb_log | env_stats
+
                 # loop through population
                 for idx, (overall_fitness, gen_fitness) in enumerate(
                     zip(log["top_fitness"], log["top_gen_fitness"])
