+++ pax/watchers.py
@@ -374,53 +374,3 @@ class ESLog(object):
         ax.spines["top"].set_visible(False)
         ax.spines["right"].set_visible(False)
         return fig, ax
-
-
-def ipd_visitation(traj, final_t) -> dict:
-    # obs [num_outer_steps, num_inner_steps, num_opps, num_envs, ...]
-    # final_t [num_opps, num_envs, ...]
-    num_timesteps = traj.observations.shape[0] * traj.observations.shape[1]
-    # obs = [0, 1, 2, 3, 4], a = [0, 1]
-    # combine = [0, .... 9]
-    state_actions = 2 * jnp.argmax(traj.observations, axis=-1) + traj.actions
-    state_actions = jnp.reshape(
-        state_actions,
-        (num_timesteps,) + state_actions.shape[2:],
-    )
-    # assume final step taken is cooperate
-    final_obs = jax.lax.expand_dims(
-        2 * jnp.argmax(final_t.observation, axis=-1), [0]
-    )
-    state_actions = jnp.append(state_actions, final_obs, axis=0)
-    hist = jnp.bincount(state_actions.flatten(), length=10)
-    state_freq = hist.reshape((int(hist.shape[0] / 2), 2)).sum(axis=1)
-    action_probs = jnp.nan_to_num(hist[::2] / state_freq)
-    return {
-        "train/state_visitation/CC": state_freq[0],
-        "train/state_visitation/CD": state_freq[1],
-        "train/state_visitation/DC": state_freq[2],
-        "train/state_visitation/DD": state_freq[3],
-        "train/state_visitation/START": state_freq[4],
-        "train/cooperation_probability/CC": action_probs[0],
-        "train/cooperation_probability/CD": action_probs[1],
-        "train/cooperation_probability/DC": action_probs[2],
-        "train/cooperation_probability/DD": action_probs[3],
-        "train/cooperation_probability/START": action_probs[4],
-    }
-
-
-def cg_visitation(traj1, traj2) -> dict:
-    defect_1 = (traj1.rewards == -2).sum()
-    defect_2 = (traj2.rewards == -2).sum()
-
-    total_1 = (traj1.rewards == 1).sum()
-    total_2 = (traj2.rewards == 1).sum()
-
-    prob_1 = defect_2 / total_1
-    prob_2 = defect_1 / total_2
-    return {
-        "train/prob_coop/1": 1 - prob_1,
-        "train/prob_coop/2": 1 - prob_2,
-        "train/total_coins/1": total_1 / jnp.size(traj1.rewards),
-        "train/total_coins/2": total_2 / jnp.size(traj2.rewards),
-    }
